+++ meridian/model/posterior_sampler_test.py
@@ -61,7 +61,8 @@ class PosteriorMCMCSamplerTest(
             sigma=tfp.distributions.Deterministic(0),
             roi_m=tfp.distributions.Deterministic(0),
             roi_rf=tfp.distributions.Deterministic(0),
-        )
+        ),
+        media_effects_dist=constants.MEDIA_EFFECTS_NORMAL,
     )
     meridian = model.Meridian(
         input_data=self.short_input_data_with_media_and_rf,