+++ CHANGELOG.md
@@ -28,6 +28,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
   support.
 * Raise an error when `kpi_scaled` is all zero and `paid_media_prior_type` is
   anything other than "coefficient".
+* Prevent negative media effect priors when using lognormal distribution.
 
 ## [1.0.8] - 2025-04-08